requirements.txt

adafruit-circuitpython-rgb-display
mysql-connector-python
pn532pi
pyserial
Pillow
flask
Flask-QRcode
pyotp
